+++ examples/send_e2e_image.rs
@@ -49,10 +49,7 @@ async fn main() {
         println!("Could not fetch public key: {}", e);
         process::exit(1);
     });
-    let recipient_key: RecipientKey = public_key.parse().unwrap_or_else(|e| {
-        println!("{}", e);
-        process::exit(1);
-    });
+    let recipient_key: RecipientKey = public_key.into();
 
     // Encrypt image
     let mut file = File::open(path).unwrap_or_else(|e| {